requirements/base.txt
Django
psycopg2
psycopg2-binary
django-versatileimagefield
django-jet
djangorestframework
pyresttest
django-cors-headers
django-froala-editor